refactor: refactor MainPage and BookDetailSection
