+++ src/components/books-cards/BookCard.module.scss
@@ -63,7 +63,7 @@
   text-align: center;
 }
 
-@media (max-width: 800px) {
+@media (max-width: 900px) {
   .image {
     display: flex;
     width: 100%;
