+++ src/pages/book-detail-section/BookDetailSection.module.scss
@@ -13,3 +13,8 @@
   height: 1px;
   width: 100%;
 }
+
+.error {
+  color: var(--clr-error-1);
+  margin-top: 100px;
+}
