+++ src/pages/main-page/components/search-section/SearchSection.module.scss
@@ -14,7 +14,7 @@
   width: 90%;
 }
 
-@media (max-width: 800px) {
+@media (max-width: 950px) {
   .input {
     width: 100%;
   }
